rename useColorPalette.ts into useThemeColorPalette.ts, remove default export
