+++ apps/src/shared/ThemedText.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Text, TextProps } from 'react-native';
-import useThemeColorPalette from './styling/theme/useColorPalette';
+import { useThemeColorPalette } from './styling/theme/useThemeColorPalette';
 
 export const ThemedText = ({ children, style, ...props }: TextProps) => {
   const { colors } = useThemeColorPalette();
