+++ apps/src/shared/ThemedView.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { View, ViewProps } from 'react-native';
-import useThemeColorPalette from './styling/theme/useColorPalette';
+import { useThemeColorPalette } from './styling/theme/useThemeColorPalette';
 
 export const ThemedView = ({ children, style, ...props }: ViewProps) => {
   const { colors } = useThemeColorPalette();
